feat: Add opengraph image for link previews. Update metadata. Refactor layout to use internal client component. Shift client logic to LayoutClientComponent.
